feature: cache binary parsers to reduce parsing time
The binary parser library generates some javascript code based on the
configuration we give it and then parses that code. This is a very expensive
operation.

We recreated the parsers every time for each individual value. By reusing
them we save a lot of time.

This simple changes improves performance by about 20 times in my tests.

We have to keep a separate parser for each different length we can get
because changing the length option requires recompilation. In my
tests we only need 20 different parsers for a 7GB gopro hero8 file.